Implement structure for unit description
This commit adds a new structure UnitDescription which builds on top of
the previously imlemented collections IDs. It contains all the necessary
information (excluding the collection ID for now) for querying and
correctly converting measurements from the module. Along with it comes
a function returning measurement values according to the passed
UnitDescription, as well as some pre-defined structures for measurements
which may be commonly used.