Add MQTT functionality to the program
This commit adds MQTT functionality using the Paho MQTT asynchronous
library. So far it manages to create a client and connect it to
a broker, also featuring automatic reconnects on connection loss. It
seems as if not all the callback functions are working properly, which
should be looked into at some later time.