Implement calculation the process images size
This commit is a first step towards dynamically allocating the process
images depending on the module configuration on the bus. This
feature uses the KBus DBus interface to retrieve information on
the connected modules. Input/output sizes are calculated using the new
function get_process_data_size. The output is only informational for
now.
Some more significant changes:
- Retrieving the KBus information requires the application to be in
  'unconfigured' state, so one additional state change including all the
  error checking is required. To remove redundancy, a new function
  set_application_state has been added. For the same reason,
  find_and_initialize_kbus does no longer set the application state to
  'running' itself.
- Since there are now even more function calls which may result in an
  error, most of which are handled the same way a new macro
  exit_on_failure has been introduced to reduce some repetition in the
  code.