Dynamically allocate all process data addresses
This commit adds a new function get_pm_data_addresses which searches the
KBus for all connected 495/494 power measurement modules and returns
their memory addresses within the process image data as an array of
Type595ProcessInput/Output pointers, along with the number of found
modules. This allows to dynamically scale the application to whichever
number of modules is found on the bus.