Prepare for sending multiple ResultSets
This commit adds a new member to ResultSet containing the index of the
power measurement module, and a new function to allocate an array of
ResultSets according to the number of modules found on the bus. The
index is also included in the MQTT message string, however, the current
code still only works with the first result. This should be the final
step in preparing for working with multiple modules.