Process every module in the main loop
This commit finishes the implementation of support for multiple power
measurement modules by iterating and processing the process data for
each single one. While it should work in theory, however, it could not
actually be verified due to the lack of more than one module on the test
system available.
The console output of measurement data has been dropped with this
commit, as maintaining it for multiple modules, each of which may or may
not have a complete set of measurements a given time, has been
considered too much of a hassle. Viewing the measurements thus has to be
done through a MQTT client from now on.