Use Protocol Buffers for measurement results
This rather huge commit introduces Protocol Buffers for a more efficient
transmission of the measurement results. It includes a modified makefile
for protobof updated to the latest version, and two new rule files for
the Protobuf C library, as well as a .proto definition file of our
message format together with the according .h/.c files. The Makefile as
well as the deploy script have also been updated accordingly.