Fix time overflow bug
On some occasions the value for runtimeUs would overflow. This happens
because the seconds and nanoseconds parts of startTime and finishTime
were subtracted separately, producing an overflow if the nanoseconds
part of startTime happened to be bigger than the one of finishTime. This
commit fixes the bug by first calculating the microseconds of both the
start and finish time and only then subtracting them.